requirements.txt
beautifulsoup4==4.4.0
html5lib==0.99999
jsbeautifier==1.5.10
pefile==1.2.10-114
lxml==3.4.4
chardet==2.3.0
requests==2.6.0
PySocks==1.5.6
cssutils==1.0
zope.interface==4.1.2
pyparsing==2.0.3
python-magic==0.4.6
rarfile==2.7
networkx==1.10
pymongo==3.0.1
ssdeep==3.1
six==1.9.0
pika==0.9.14
yara==1.7.6

